♻️ (dashboard/alerts): Move applyClientFilters into its own module
Co-locating the pure helper with the `<AlertFilterBar>` component
breaks Vite's React fast-refresh rule (a file may export either
components or constants/functions, not both). Extracting it into
`alertFilters.ts` clears the warning without changing call sites
beyond their import paths.
