+++ dashboard/src/features/alerts/AlertFilterBar.tsx
@@ -156,18 +156,3 @@ export function AlertFilterBar({ value, onChange }: AlertFilterBarProps) {
   )
 }
 
-export function applyClientFilters(
-  rows: readonly import('./types').Alert[],
-  filters: AlertFilters,
-): readonly import('./types').Alert[] {
-  const q = filters.agentQuery.trim().toLowerCase()
-  return rows.filter((r) => {
-    if (filters.severities.length && !filters.severities.includes(r.severity)) return false
-    if (filters.statuses.length && !filters.statuses.includes(r.status)) return false
-    if (q) {
-      const haystack = `${r.agentId ?? ''} ${r.ruleName}`.toLowerCase()
-      if (!haystack.includes(q)) return false
-    }
-    return true
-  })
-}
